tc-53 fjerna useFetch og bytte til useQuery
