+++ frontend/mulighetsrommet-veileder-flate/src/api/utils.ts
@@ -1,12 +1,4 @@
-import { hasData as _hasData, WithData, WithError, WithoutData } from '@nutgaard/use-fetch';
 import { APPLICATION_NAME } from '../constants';
-
-type AsyncData<TYPE> = WithoutData | WithData<TYPE> | WithError;
-
-// Midlertidig fix for at hasData fra useFetch sjekker status og ikke at data er lastet inn
-export function hasData<TYPE>(result: AsyncData<TYPE>): result is WithData<TYPE> {
-  return (result as WithData<TYPE>).data != null && _hasData(result);
-}
 
 export const headers = {
   'Nav-Consumer-Id': APPLICATION_NAME,
