+++ frontend/mulighetsrommet-veileder-flate/src/mock/api/features.ts
@@ -1,4 +1,4 @@
-import { Features, ALERT_INFO, FAKE_DOOR } from '../../api/features';
+import { Features, ALERT_INFO, FAKE_DOOR } from '../../api/feature-toggles';
 
 export const mockFeatures: Features = {
   [ALERT_INFO]: true,
